refactored studForm for readability
