+++ front-end/src/App.tsx
@@ -2,13 +2,10 @@ import "./index.css";
 
 import { Route, BrowserRouter as Router, Switch } from "react-router-dom";
 
-import { Grid } from "@material-ui/core";
 import HomeContainer from "./containers/HomeContainer";
 import LoginForm from "./components/LoginForm";
 import LoginPageContainer from "./containers/LoginPageContainer/index";
-import React from "react";
 import RegisterForm from "./components/RegisterForm";
-import StudForm from "./components/StudForm";
 import { useStudUser } from "./hooks/useStudUser";
 
 const App = () => {
