+++ front-end/src/components/StudTextField/index.css
@@ -1,12 +1,12 @@
-.login-card .input-field-container .form-title {
-	color: white;
-	padding-bottom: 10px;
-  }
-  
-  .login-card .input-field-container .input-field-item {
-	padding: 10px;
-  }
-  
-  .login-card .input-field-container .input-field-item .input {
-	background-color: white;
-  }
+.form-card .input-field-container .form-title {
+  color: white;
+  padding-bottom: 10px;
+}
+
+.form-card .input-field-container .input-field-item {
+  padding: 10px;
+}
+
+.form-card .input-field-container .input-field-item .input {
+  background-color: white;
+}
