+++ front-end/src/types/formTypes.ts
@@ -1,0 +1,8 @@
+// fields that should be passed into formField
+export type FormFieldType = {
+  fieldTitle: string;
+  fieldType: "textField" | "dropDown" | "imageUploader";
+  handleOnChange: Function;
+  inputTypePassword?: boolean;
+  selectItems?: Array<string>;
+};
